+++ src/Exception/LogicException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Automate\Exception;
+
+class LogicException extends CustomException
+{
+}