+++ tests/unit/Transformer/Logic/MyLogic.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Automate\Transformer\Logic;
+
+class MyLogic extends Logic
+{
+    public function answeredBy() : bool
+    {
+        return true;
+    }
+}